adaptix/services/intelligence/requirements.txt
django==4.2
djangorestframework
django-cors-headers
dj-database-url
psycopg2-binary
gunicorn
python-dotenv
# Machine Learning & Data Science
numpy
pandas
scikit-learn
prophet
redis
celery
kombu
pytest==8.0.0
pytest-django==4.8.0
pytest-mock==3.12.0
drf-spectacular
cryptography
djangorestframework-simplejwt
pika==1.3.2
